requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
streamlit>=1.47.0
opencv-python>=4.10.0
numpy>=1.26.0
ultralytics>=8.3.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
braintrust>=0.5.6
python-multipart>=0.0.9
